9ns: Use an int for perm and mode
-1 means "don't change the mode."  That was 0x00000000ffffffff, which is
not ~0UL (0xffffffffffffffff).  Just use an int and don't play games with
the sizeof longs and whatnot.

+++ kern/include/ns.h
@@ -139,7 +139,7 @@ struct dir {
 	unsigned int dev;			/* server subtype */
 	/* file data */
 	struct qid qid;				/* unique id from server */
-	uint32_t mode;				/* permissions */
+	int mode;					/* permissions */
 	uint32_t atime;				/* last read time */
 	uint32_t mtime;				/* last write time */
 	int64_t length;				/* file length: see <u.h> */
@@ -475,7 +475,7 @@ struct dirtab {
 	char name[KNAMELEN];
 	struct qid qid;
 	int64_t length;
-	long perm;
+	int perm;
 };
 
 struct walkqid {
